requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ravendb==7.1.3
python-telegram-bot==20.7
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
